Change WasmMemory.byteSize() return type to long.

+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/memory/ByteArrayWasmMemory.java
@@ -128,7 +128,7 @@ public final class ByteArrayWasmMemory extends WasmMemory {
     }
 
     @Override
-    public int byteSize() {
+    public long byteSize() {
         return buffer.length;
     }
